weather app/requirements.txt
Django==4.2.7
python-decouple==3.8
requests==2.31.0
